backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# PDF Processing
PyPDF2==3.0.1

# Data & Models
pydantic==2.6.1
numpy==1.26.4
Pillow==10.2.0

# LLM & Image Generation
openai==1.14.2
transformers==4.39.3
diffusers==0.27.2
accelerate==0.28.0
sentencepiece==0.2.0
safetensors==0.4.2

# Animation / Video
moviepy==1.0.3
imageio==2.34.0
imageio-ffmpeg==0.5.1

# Utilities
requests==2.31.0
tqdm==4.66.2

# Optional: For GPU
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
